modularized the if statements for the drag and drop and state resetting
